scheduler: clusterInfo refactor
move storesInfo, regionsInfo and regionTree into core
