+++ server/cache.go
@@ -14,7 +14,6 @@
 package server
 
 import (
-	"math/rand"
 	"sync"
 	"time"
 
@@ -28,12 +27,6 @@ import (
 )
 
 var (
-	errStoreNotFound = func(storeID uint64) error {
-		return errors.Errorf("store %v not found", storeID)
-	}
-	errStoreIsBlocked = func(storeID uint64) error {
-		return errors.Errorf("store %v is blocked", storeID)
-	}
 	errRegionNotFound = func(regionID uint64) error {
 		return errors.Errorf("region %v not found", regionID)
 	}
@@ -42,309 +35,6 @@ var (
 	}
 )
 
-type storesInfo struct {
-	stores map[uint64]*core.StoreInfo
-}
-
-func newStoresInfo() *storesInfo {
-	return &storesInfo{
-		stores: make(map[uint64]*core.StoreInfo),
-	}
-}
-
-func (s *storesInfo) getStore(storeID uint64) *core.StoreInfo {
-	store, ok := s.stores[storeID]
-	if !ok {
-		return nil
-	}
-	return store.Clone()
-}
-
-func (s *storesInfo) setStore(store *core.StoreInfo) {
-	s.stores[store.GetId()] = store
-}
-
-func (s *storesInfo) blockStore(storeID uint64) error {
-	store, ok := s.stores[storeID]
-	if !ok {
-		return errStoreNotFound(storeID)
-	}
-	if store.IsBlocked() {
-		return errStoreIsBlocked(storeID)
-	}
-	store.Block()
-	return nil
-}
-
-func (s *storesInfo) unblockStore(storeID uint64) {
-	store, ok := s.stores[storeID]
-	if !ok {
-		log.Fatalf("store %d is unblocked, but it is not found", storeID)
-	}
-	store.Unblock()
-}
-
-func (s *storesInfo) getStores() []*core.StoreInfo {
-	stores := make([]*core.StoreInfo, 0, len(s.stores))
-	for _, store := range s.stores {
-		stores = append(stores, store.Clone())
-	}
-	return stores
-}
-
-func (s *storesInfo) getMetaStores() []*metapb.Store {
-	stores := make([]*metapb.Store, 0, len(s.stores))
-	for _, store := range s.stores {
-		stores = append(stores, proto.Clone(store.Store).(*metapb.Store))
-	}
-	return stores
-}
-
-func (s *storesInfo) getStoreCount() int {
-	return len(s.stores)
-}
-
-func (s *storesInfo) setLeaderCount(storeID uint64, leaderCount int) {
-	if store, ok := s.stores[storeID]; ok {
-		store.LeaderCount = leaderCount
-	}
-}
-
-func (s *storesInfo) setRegionCount(storeID uint64, regionCount int) {
-	if store, ok := s.stores[storeID]; ok {
-		store.RegionCount = regionCount
-	}
-}
-
-func (s *storesInfo) totalWrittenBytes() uint64 {
-	var totalWrittenBytes uint64
-	for _, s := range s.stores {
-		if s.IsUp() {
-			totalWrittenBytes += s.Stats.GetBytesWritten()
-		}
-	}
-	return totalWrittenBytes
-}
-
-func (s *storesInfo) totalReadBytes() uint64 {
-	var totalReadBytes uint64
-	for _, s := range s.stores {
-		if s.IsUp() {
-			totalReadBytes += s.Stats.GetBytesRead()
-		}
-	}
-	return totalReadBytes
-}
-
-// regionMap wraps a map[uint64]*core.RegionInfo and supports randomly pick a region.
-type regionMap struct {
-	m   map[uint64]*regionEntry
-	ids []uint64
-}
-
-type regionEntry struct {
-	*core.RegionInfo
-	pos int
-}
-
-func newRegionMap() *regionMap {
-	return &regionMap{
-		m: make(map[uint64]*regionEntry),
-	}
-}
-
-func (rm *regionMap) Len() int {
-	if rm == nil {
-		return 0
-	}
-	return len(rm.m)
-}
-
-func (rm *regionMap) Get(id uint64) *core.RegionInfo {
-	if rm == nil {
-		return nil
-	}
-	if entry, ok := rm.m[id]; ok {
-		return entry.RegionInfo
-	}
-	return nil
-}
-
-func (rm *regionMap) Put(region *core.RegionInfo) {
-	if old, ok := rm.m[region.GetId()]; ok {
-		old.RegionInfo = region
-		return
-	}
-	rm.m[region.GetId()] = &regionEntry{
-		RegionInfo: region,
-		pos:        len(rm.ids),
-	}
-	rm.ids = append(rm.ids, region.GetId())
-}
-
-func (rm *regionMap) RandomRegion() *core.RegionInfo {
-	if rm.Len() == 0 {
-		return nil
-	}
-	return rm.Get(rm.ids[rand.Intn(rm.Len())])
-}
-
-func (rm *regionMap) Delete(id uint64) {
-	if rm == nil {
-		return
-	}
-	if old, ok := rm.m[id]; ok {
-		len := rm.Len()
-		last := rm.m[rm.ids[len-1]]
-		last.pos = old.pos
-		rm.ids[last.pos] = last.GetId()
-		delete(rm.m, id)
-		rm.ids = rm.ids[:len-1]
-	}
-}
-
-type regionsInfo struct {
-	tree      *regionTree
-	regions   *regionMap            // regionID -> regionInfo
-	leaders   map[uint64]*regionMap // storeID -> regionID -> regionInfo
-	followers map[uint64]*regionMap // storeID -> regionID -> regionInfo
-}
-
-func newRegionsInfo() *regionsInfo {
-	return &regionsInfo{
-		tree:      newRegionTree(),
-		regions:   newRegionMap(),
-		leaders:   make(map[uint64]*regionMap),
-		followers: make(map[uint64]*regionMap),
-	}
-}
-
-func (r *regionsInfo) getRegion(regionID uint64) *core.RegionInfo {
-	region := r.regions.Get(regionID)
-	if region == nil {
-		return nil
-	}
-	return region.Clone()
-}
-
-func (r *regionsInfo) setRegion(region *core.RegionInfo) {
-	if origin := r.regions.Get(region.GetId()); origin != nil {
-		r.removeRegion(origin)
-	}
-	r.addRegion(region)
-}
-
-func (r *regionsInfo) addRegion(region *core.RegionInfo) {
-	// Add to tree and regions.
-	r.tree.update(region.Region)
-	r.regions.Put(region)
-
-	if region.Leader == nil {
-		return
-	}
-
-	// Add to leaders and followers.
-	for _, peer := range region.GetPeers() {
-		storeID := peer.GetStoreId()
-		if peer.GetId() == region.Leader.GetId() {
-			// Add leader peer to leaders.
-			store, ok := r.leaders[storeID]
-			if !ok {
-				store = newRegionMap()
-				r.leaders[storeID] = store
-			}
-			store.Put(region)
-		} else {
-			// Add follower peer to followers.
-			store, ok := r.followers[storeID]
-			if !ok {
-				store = newRegionMap()
-				r.followers[storeID] = store
-			}
-			store.Put(region)
-		}
-	}
-}
-
-func (r *regionsInfo) removeRegion(region *core.RegionInfo) {
-	// Remove from tree and regions.
-	r.tree.remove(region.Region)
-	r.regions.Delete(region.GetId())
-
-	// Remove from leaders and followers.
-	for _, peer := range region.GetPeers() {
-		storeID := peer.GetStoreId()
-		r.leaders[storeID].Delete(region.GetId())
-		r.followers[storeID].Delete(region.GetId())
-	}
-}
-
-func (r *regionsInfo) searchRegion(regionKey []byte) *core.RegionInfo {
-	region := r.tree.search(regionKey)
-	if region == nil {
-		return nil
-	}
-	return r.getRegion(region.GetId())
-}
-
-func (r *regionsInfo) getRegions() []*core.RegionInfo {
-	regions := make([]*core.RegionInfo, 0, r.regions.Len())
-	for _, region := range r.regions.m {
-		regions = append(regions, region.Clone())
-	}
-	return regions
-}
-
-func (r *regionsInfo) getMetaRegions() []*metapb.Region {
-	regions := make([]*metapb.Region, 0, r.regions.Len())
-	for _, region := range r.regions.m {
-		regions = append(regions, proto.Clone(region.Region).(*metapb.Region))
-	}
-	return regions
-}
-
-func (r *regionsInfo) getRegionCount() int {
-	return r.regions.Len()
-}
-
-func (r *regionsInfo) getStoreRegionCount(storeID uint64) int {
-	return r.getStoreLeaderCount(storeID) + r.getStoreFollowerCount(storeID)
-}
-
-func (r *regionsInfo) getStoreLeaderCount(storeID uint64) int {
-	return r.leaders[storeID].Len()
-}
-
-func (r *regionsInfo) getStoreFollowerCount(storeID uint64) int {
-	return r.followers[storeID].Len()
-}
-
-func (r *regionsInfo) randRegion() *core.RegionInfo {
-	return randRegion(r.regions)
-}
-
-func (r *regionsInfo) randLeaderRegion(storeID uint64) *core.RegionInfo {
-	return randRegion(r.leaders[storeID])
-}
-
-func (r *regionsInfo) randFollowerRegion(storeID uint64) *core.RegionInfo {
-	return randRegion(r.followers[storeID])
-}
-
-const randomRegionMaxRetry = 10
-
-func randRegion(regions *regionMap) *core.RegionInfo {
-	for i := 0; i < randomRegionMaxRetry; i++ {
-		region := regions.RandomRegion()
-		if region == nil {
-			return nil
-		}
-		if len(region.DownPeers) == 0 && len(region.PendingPeers) == 0 {
-			return region.Clone()
-		}
-	}
-	return nil
-}
 
 type clusterInfo struct {
 	sync.RWMutex
@@ -352,8 +42,8 @@ type clusterInfo struct {
 	id      IDAllocator
 	kv      *kv
 	meta    *metapb.Cluster
-	stores  *storesInfo
-	regions *regionsInfo
+	stores  *core.StoresInfo
+	regions *core.RegionsInfo
 
 	activeRegions   int
 	writeStatistics cache.Cache
@@ -363,8 +53,8 @@ type clusterInfo struct {
 func newClusterInfo(id IDAllocator) *clusterInfo {
 	return &clusterInfo{
 		id:              id,
-		stores:          newStoresInfo(),
-		regions:         newRegionsInfo(),
+		stores:          core.NewStoresInfo(),
+		regions:         core.NewRegionsInfo(),
 		readStatistics:  cache.NewCache(statCacheMaxLen, cache.TwoQueueCache),
 		writeStatistics: cache.NewCache(statCacheMaxLen, cache.TwoQueueCache),
 	}
@@ -388,13 +78,13 @@ func loadClusterInfo(id IDAllocator, kv *kv) (*clusterInfo, error) {
 	if err := kv.loadStores(c.stores, kvRangeLimit); err != nil {
 		return nil, errors.Trace(err)
 	}
-	log.Infof("load %v stores cost %v", c.stores.getStoreCount(), time.Since(start))
+	log.Infof("load %v stores cost %v", c.stores.GetStoreCount(), time.Since(start))
 
 	start = time.Now()
 	if err := kv.loadRegions(c.regions, kvRangeLimit); err != nil {
 		return nil, errors.Trace(err)
 	}
-	log.Infof("load %v regions cost %v", c.regions.getRegionCount(), time.Since(start))
+	log.Infof("load %v regions cost %v", c.regions.GetRegionCount(), time.Since(start))
 
 	return c, nil
 }
@@ -449,7 +139,7 @@ func (c *clusterInfo) putMetaLocked(meta *metapb.Cluster) error {
 func (c *clusterInfo) GetStore(storeID uint64) *core.StoreInfo {
 	c.RLock()
 	defer c.RUnlock()
-	return c.stores.getStore(storeID)
+	return c.stores.GetStore(storeID)
 }
 
 func (c *clusterInfo) putStore(store *core.StoreInfo) error {
@@ -464,7 +154,7 @@ func (c *clusterInfo) putStoreLocked(store *core.StoreInfo) error {
 			return errors.Trace(err)
 		}
 	}
-	c.stores.setStore(store)
+	c.stores.SetStore(store)
 	return nil
 }
 
@@ -472,60 +162,52 @@ func (c *clusterInfo) putStoreLocked(store *core.StoreInfo) error {
 func (c *clusterInfo) BlockStore(storeID uint64) error {
 	c.Lock()
 	defer c.Unlock()
-	return errors.Trace(c.stores.blockStore(storeID))
+	return errors.Trace(c.stores.BlockStore(storeID))
 }
 
 // UnblockStore allows balancer to select the store.
 func (c *clusterInfo) UnblockStore(storeID uint64) {
 	c.Lock()
 	defer c.Unlock()
-	c.stores.unblockStore(storeID)
+	c.stores.UnblockStore(storeID)
 }
 
 // GetStores returns all stores in the cluster.
 func (c *clusterInfo) GetStores() []*core.StoreInfo {
 	c.RLock()
 	defer c.RUnlock()
-	return c.stores.getStores()
+	return c.stores.GetStores()
 }
 
 func (c *clusterInfo) getMetaStores() []*metapb.Store {
 	c.RLock()
 	defer c.RUnlock()
-	return c.stores.getMetaStores()
+	return c.stores.GetMetaStores()
 }
 
 func (c *clusterInfo) getStoreCount() int {
 	c.RLock()
 	defer c.RUnlock()
-	return c.stores.getStoreCount()
+	return c.stores.GetStoreCount()
 }
 
 func (c *clusterInfo) getStoresWriteStat() map[uint64]uint64 {
 	c.RLock()
 	defer c.RUnlock()
-	res := make(map[uint64]uint64)
-	for _, s := range c.stores.stores {
-		res[s.GetId()] = s.Stats.GetBytesWritten()
-	}
-	return res
+	return c.stores.GetStoresWriteStat()
 }
 
 func (c *clusterInfo) getStoresReadStat() map[uint64]uint64 {
 	c.RLock()
 	defer c.RUnlock()
-	res := make(map[uint64]uint64)
-	for _, s := range c.stores.stores {
-		res[s.GetId()] = s.Stats.GetBytesRead()
-	}
-	return res
+    return c.stores.GetStoresReadStat()
 }
 
 // GetRegions searches for a region by ID.
 func (c *clusterInfo) GetRegion(regionID uint64) *core.RegionInfo {
 	c.RLock()
 	defer c.RUnlock()
-	return c.regions.getRegion(regionID)
+	return c.regions.GetRegion(regionID)
 }
 
 // updateWriteStatCache updates statistic for a region if it's hot, or remove it from statistics if it cools down
@@ -631,7 +313,7 @@ func (c *clusterInfo) IsRegionHot(id uint64) bool {
 func (c *clusterInfo) searchRegion(regionKey []byte) *core.RegionInfo {
 	c.RLock()
 	defer c.RUnlock()
-	return c.regions.searchRegion(regionKey)
+	return c.regions.SearchRegion(regionKey)
 }
 
 func (c *clusterInfo) putRegion(region *core.RegionInfo) error {
@@ -646,58 +328,58 @@ func (c *clusterInfo) putRegionLocked(region *core.RegionInfo) error {
 			return errors.Trace(err)
 		}
 	}
-	c.regions.setRegion(region)
+	c.regions.SetRegion(region)
 	return nil
 }
 
 func (c *clusterInfo) getRegions() []*core.RegionInfo {
 	c.RLock()
 	defer c.RUnlock()
-	return c.regions.getRegions()
+	return c.regions.GetRegions()
 }
 
 func (c *clusterInfo) randomRegion() *core.RegionInfo {
 	c.RLock()
 	defer c.RUnlock()
-	return c.regions.randRegion()
+	return c.regions.RandRegion()
 }
 
 func (c *clusterInfo) getMetaRegions() []*metapb.Region {
 	c.RLock()
 	defer c.RUnlock()
-	return c.regions.getMetaRegions()
+	return c.regions.GetMetaRegions()
 }
 
 func (c *clusterInfo) getRegionCount() int {
 	c.RLock()
 	defer c.RUnlock()
-	return c.regions.getRegionCount()
+	return c.regions.GetRegionCount()
 }
 
 func (c *clusterInfo) getStoreRegionCount(storeID uint64) int {
 	c.RLock()
 	defer c.RUnlock()
-	return c.regions.getStoreRegionCount(storeID)
+	return c.regions.GetStoreRegionCount(storeID)
 }
 
 func (c *clusterInfo) getStoreLeaderCount(storeID uint64) int {
 	c.RLock()
 	defer c.RUnlock()
-	return c.regions.getStoreLeaderCount(storeID)
+	return c.regions.GetStoreLeaderCount(storeID)
 }
 
 // RandLeaderRegion returns a random region that has leader on the store.
 func (c *clusterInfo) RandLeaderRegion(storeID uint64) *core.RegionInfo {
 	c.RLock()
 	defer c.RUnlock()
-	return c.regions.randLeaderRegion(storeID)
+	return c.regions.RandLeaderRegion(storeID)
 }
 
 // RandFolowerRegion returns a random region that has a follower on the store.
 func (c *clusterInfo) RandFollowerRegion(storeID uint64) *core.RegionInfo {
 	c.RLock()
 	defer c.RUnlock()
-	return c.regions.randFollowerRegion(storeID)
+	return c.regions.RandFollowerRegion(storeID)
 }
 
 // GetRegionStores returns all stores that contains the region's peer.
@@ -706,7 +388,7 @@ func (c *clusterInfo) GetRegionStores(region *core.RegionInfo) []*core.StoreInfo
 	defer c.RUnlock()
 	var stores []*core.StoreInfo
 	for id := range region.GetStoreIds() {
-		if store := c.stores.getStore(id); store != nil {
+		if store := c.stores.GetStore(id); store != nil {
 			stores = append(stores, store)
 		}
 	}
@@ -717,7 +399,7 @@ func (c *clusterInfo) GetRegionStores(region *core.RegionInfo) []*core.StoreInfo
 func (c *clusterInfo) GetLeaderStore(region *core.RegionInfo) *core.StoreInfo {
 	c.RLock()
 	defer c.RUnlock()
-	return c.stores.getStore(region.Leader.GetStoreId())
+	return c.stores.GetStore(region.Leader.GetStoreId())
 }
 
 // GetRegionStores returns all stores that contains the region's follower peer.
@@ -726,7 +408,7 @@ func (c *clusterInfo) GetFollowerStores(region *core.RegionInfo) []*core.StoreIn
 	defer c.RUnlock()
 	var stores []*core.StoreInfo
 	for id := range region.GetFollowers() {
-		if store := c.stores.getStore(id); store != nil {
+		if store := c.stores.GetStore(id); store != nil {
 			stores = append(stores, store)
 		}
 	}
@@ -737,7 +419,7 @@ func (c *clusterInfo) GetFollowerStores(region *core.RegionInfo) []*core.StoreIn
 func (c *clusterInfo) isPrepared() bool {
 	c.RLock()
 	defer c.RUnlock()
-	return float64(c.regions.regions.Len())*collectFactor <= float64(c.activeRegions)
+	return float64(c.regions.Length())*collectFactor <= float64(c.activeRegions)
 }
 
 // handleStoreHeartbeat updates the store status.
@@ -746,27 +428,27 @@ func (c *clusterInfo) handleStoreHeartbeat(stats *pdpb.StoreStats) error {
 	defer c.Unlock()
 
 	storeID := stats.GetStoreId()
-	store := c.stores.getStore(storeID)
+	store := c.stores.GetStore(storeID)
 	if store == nil {
-		return errors.Trace(errStoreNotFound(storeID))
+		return errors.Trace(core.ErrStoreNotFound(storeID))
 	}
 	store.Stats = proto.Clone(stats).(*pdpb.StoreStats)
 	store.LastHeartbeatTS = time.Now()
 
-	c.stores.setStore(store)
+	c.stores.SetStore(store)
 	return nil
 }
 
 func (c *clusterInfo) updateStoreStatus(id uint64) {
-	c.stores.setLeaderCount(id, c.regions.getStoreLeaderCount(id))
-	c.stores.setRegionCount(id, c.regions.getStoreRegionCount(id))
+	c.stores.SetLeaderCount(id, c.regions.GetStoreLeaderCount(id))
+	c.stores.SetRegionCount(id, c.regions.GetStoreRegionCount(id))
 }
 
 // handleRegionHeartbeat updates the region information.
 func (c *clusterInfo) handleRegionHeartbeat(region *core.RegionInfo) error {
 	region = region.Clone()
 	c.RLock()
-	origin := c.regions.getRegion(region.GetId())
+	origin := c.regions.GetRegion(region.GetId())
 	c.RUnlock()
 
 	// Save to KV if meta is updated.
@@ -784,11 +466,11 @@ func (c *clusterInfo) handleRegionHeartbeat(region *core.RegionInfo) error {
 			return errors.Trace(errRegionIsStale(region.Region, origin.Region))
 		}
 		if r.GetVersion() > o.GetVersion() {
-			log.Infof("[region %d] %s, Version changed from {%d} to {%d}", region.GetId(), diffRegionKeyInfo(origin, region), o.GetVersion(), r.GetVersion())
+			log.Infof("[region %d] %s, Version changed from {%d} to {%d}", region.GetId(), core.DiffRegionKeyInfo(origin, region), o.GetVersion(), r.GetVersion())
 			saveKV, saveCache = true, true
 		}
 		if r.GetConfVer() > o.GetConfVer() {
-			log.Infof("[region %d] %s, ConfVer changed from {%d} to {%d}", region.GetId(), diffRegionPeersInfo(origin, region), o.GetConfVer(), r.GetConfVer())
+			log.Infof("[region %d] %s, ConfVer changed from {%d} to {%d}", region.GetId(), core.DiffRegionPeersInfo(origin, region), o.GetConfVer(), r.GetConfVer())
 			saveKV, saveCache = true, true
 		}
 		if region.Leader.GetId() != origin.Leader.GetId() {
@@ -820,7 +502,7 @@ func (c *clusterInfo) handleRegionHeartbeat(region *core.RegionInfo) error {
 	}
 
 	if saveCache {
-		c.regions.setRegion(region)
+		c.regions.SetRegion(region)
 
 		// Update related stores.
 		if origin != nil {
@@ -858,7 +540,7 @@ func (c *clusterInfo) updateWriteStatus(region *core.RegionInfo) {
 	// and we use total written Bytes past storeHeartBeatReportInterval seconds to divide the number of hot regions
 	// divide 2 because the store reports data about two times than the region record write to rocksdb
 	divisor := float64(statCacheMaxLen) * 2 * storeHeartBeatReportInterval
-	hotRegionThreshold := uint64(float64(c.stores.totalWrittenBytes()) / divisor)
+	hotRegionThreshold := uint64(float64(c.stores.TotalWrittenBytes()) / divisor)
 
 	if hotRegionThreshold < hotRegionMinFlowRate {
 		hotRegionThreshold = hotRegionMinFlowRate
@@ -884,7 +566,7 @@ func (c *clusterInfo) updateReadStatus(region *core.RegionInfo) {
 	// suppose the number of the hot regions is statLRUMaxLen
 	// and we use total written Bytes past storeHeartBeatReportInterval seconds to divide the number of hot regions
 	divisor := float64(statCacheMaxLen) * storeHeartBeatReportInterval
-	hotRegionThreshold := uint64(float64(c.stores.totalReadBytes()) / divisor)
+	hotRegionThreshold := uint64(float64(c.stores.TotalReadBytes()) / divisor)
 
 	if hotRegionThreshold < hotRegionMinFlowRate {
 		hotRegionThreshold = hotRegionMinFlowRate
