+++ server/cluster.go
@@ -466,7 +466,7 @@ func (c *RaftCluster) RemoveStore(storeID uint64) error {
 
 	store := cluster.GetStore(storeID)
 	if store == nil {
-		return errors.Trace(errStoreNotFound(storeID))
+		return errors.Trace(core.ErrStoreNotFound(storeID))
 	}
 
 	// Remove an offline store should be OK, nothing to do.
@@ -495,7 +495,7 @@ func (c *RaftCluster) BuryStore(storeID uint64, force bool) error {
 
 	store := cluster.GetStore(storeID)
 	if store == nil {
-		return errors.Trace(errStoreNotFound(storeID))
+		return errors.Trace(core.ErrStoreNotFound(storeID))
 	}
 
 	// Bury a tombstone store should be OK, nothing to do.
@@ -524,7 +524,7 @@ func (c *RaftCluster) SetStoreState(storeID uint64, state metapb.StoreState) err
 
 	store := cluster.GetStore(storeID)
 	if store == nil {
-		return errors.Trace(errStoreNotFound(storeID))
+		return errors.Trace(core.ErrStoreNotFound(storeID))
 	}
 
 	store.State = state
@@ -539,7 +539,7 @@ func (c *RaftCluster) SetStoreWeight(storeID uint64, leader, region float64) err
 
 	store := c.cachedCluster.GetStore(storeID)
 	if store == nil {
-		return errors.Trace(errStoreNotFound(storeID))
+		return errors.Trace(core.ErrStoreNotFound(storeID))
 	}
 
 	if err := c.s.kv.saveStoreWeight(storeID, leader, region); err != nil {
