+++ server/core/store.go
@@ -16,6 +16,8 @@ package core
 import (
 	"time"
 
+	log "github.com/Sirupsen/logrus"
+	"github.com/juju/errors"
 	"github.com/gogo/protobuf/proto"
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/kvproto/pkg/pdpb"
@@ -223,3 +225,125 @@ type StoreHotRegionInfos struct {
 
 // StoreHotRegionsStat used to record the hot region statistics group by store
 type StoreHotRegionsStat map[uint64]*HotRegionsStat
+
+var (
+	ErrStoreNotFound = func(storeID uint64) error {
+		return errors.Errorf("store %v not found", storeID)
+	}
+	ErrStoreIsBlocked = func(storeID uint64) error {
+		return errors.Errorf("store %v is blocked", storeID)
+	}
+)
+
+// stores info
+type StoresInfo struct {
+	stores map[uint64]*StoreInfo
+}
+
+func NewStoresInfo() *StoresInfo {
+	return &StoresInfo{
+		stores: make(map[uint64]*StoreInfo),
+	}
+}
+
+func (s *StoresInfo) GetStore(storeID uint64) *StoreInfo {
+	store, ok := s.stores[storeID]
+	if !ok {
+		return nil
+	}
+	return store.Clone()
+}
+
+func (s *StoresInfo) SetStore(store *StoreInfo) {
+	s.stores[store.GetId()] = store
+}
+
+func (s *StoresInfo) BlockStore(storeID uint64) error {
+	store, ok := s.stores[storeID]
+	if !ok {
+		return ErrStoreNotFound(storeID)
+	}
+	if store.IsBlocked() {
+		return ErrStoreIsBlocked(storeID)
+	}
+	store.Block()
+	return nil
+}
+
+func (s *StoresInfo) UnblockStore(storeID uint64) {
+	store, ok := s.stores[storeID]
+	if !ok {
+		log.Fatalf("store %d is unblocked, but it is not found", storeID)
+	}
+	store.Unblock()
+}
+
+func (s *StoresInfo) GetStores() []*StoreInfo {
+	stores := make([]*StoreInfo, 0, len(s.stores))
+	for _, store := range s.stores {
+		stores = append(stores, store.Clone())
+	}
+	return stores
+}
+
+func (s *StoresInfo) GetMetaStores() []*metapb.Store {
+	stores := make([]*metapb.Store, 0, len(s.stores))
+	for _, store := range s.stores {
+		stores = append(stores, proto.Clone(store.Store).(*metapb.Store))
+	}
+	return stores
+}
+
+func (s *StoresInfo) GetStoreCount() int {
+	return len(s.stores)
+}
+
+func (s *StoresInfo) SetLeaderCount(storeID uint64, leaderCount int) {
+	if store, ok := s.stores[storeID]; ok {
+		store.LeaderCount = leaderCount
+	}
+}
+
+func (s *StoresInfo) SetRegionCount(storeID uint64, regionCount int) {
+	if store, ok := s.stores[storeID]; ok {
+		store.RegionCount = regionCount
+	}
+}
+
+func (s *StoresInfo) TotalWrittenBytes() uint64 {
+	var totalWrittenBytes uint64
+	for _, s := range s.stores {
+		if s.IsUp() {
+			totalWrittenBytes += s.Stats.GetBytesWritten()
+		}
+	}
+	return totalWrittenBytes
+}
+
+func (s *StoresInfo) TotalReadBytes() uint64 {
+	var totalReadBytes uint64
+	for _, s := range s.stores {
+		if s.IsUp() {
+			totalReadBytes += s.Stats.GetBytesRead()
+		}
+	}
+	return totalReadBytes
+}
+
+func (s* StoresInfo) GetStoresWriteStat() map[uint64]uint64 {
+	res := make(map[uint64]uint64)
+	for _, s := range s.stores {
+		res[s.GetId()] = s.Stats.GetBytesWritten()
+	}
+    return res
+}
+
+func (s* StoresInfo) GetStoresReadStat() map[uint64]uint64 {
+	res := make(map[uint64]uint64)
+	for _, s := range s.stores {
+		res[s.GetId()] = s.Stats.GetBytesRead()
+	}
+    return res
+}
+
+
